Updated Text
Just updated some of the URL links to include them opening up on a new window. 

Also I made the last HHS.gov on the "stay updated" an a tag, so styling may change for that. Thanks for making the other updates as we discussed in the meeting! Very excited to show them.